Rename diesel.rs -> extras.rs
